#19 ♻️: refactor 커뮤니티 조회 fetch join
