+++ coinkiri-api/src/main/kotlin/coinkiri/api/controller/comment/dto/response/CommentResponseDto.kt
@@ -1,9 +1,13 @@
 package coinkiri.api.controller.comment.dto.response
 
 import coinkiri.api.controller.member.dto.response.MemberInfoDto
+import coinkiri.api.controller.post.dto.response.PostResponseDto
+import java.time.LocalDateTime
 
 class CommentResponseDto (
-    val parentComment: CommentResponseDto?,
+    val id: Long,
+    val post: PostResponseDto,
     val member: MemberInfoDto,
     val content: String,
+    val createdAt: LocalDateTime,
 )
