--- coinkiri-core/src/main/kotlin/coinkiri/core/domain/post/repository/CommunityRepository.java
@@ -1,9 +1,0 @@
-package coinkiri.core.domain.post.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import coinkiri.core.domain.post.Community;
-
-
-public interface CommunityRepository extends JpaRepository<Community, Long>{
-}
