+++ coinkiri-core/src/main/kotlin/coinkiri/core/domain/post/repository/community/CommunityRepositoryDsl.java
@@ -1,0 +1,5 @@
+package coinkiri.core.domain.post.repository.community;
+
+
+public interface CommunityRepositoryDsl {
+}
