+++ coinkiri-core/src/main/kotlin/coinkiri/core/domain/post/repository/community/CommunityRepositoryImpl.java
@@ -1,0 +1,12 @@
+package coinkiri.core.domain.post.repository.community;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class CommunityRepositoryImpl {
+	private final JPAQueryFactory queryFactory;
+
+
+}
